Add files via upload
Mediator Design pattern: is one of behavioral design pattern which provides centralized communication medium 
between differents objects in the system. If object are directly interact with one another 
system components are tightly-coupled with each other that makes higher cost of maintainability and not hard to extend.


There are 3 key components of the Mediator pattern:

Mediator: Defines the interface for communication between Colleague objects. It typically contains methods to send and receive messages and maintain references to the Colleague objects.

ConcreteMediator: Implements the Mediator interface and manages the communication and interactions between Colleague objects. It contains the logic to route messages and coordinate actions between the Colleague objects.

Colleague: Represents individual objects that need to communicate with each other. Colleague objects interact with the Mediator to send and receive messages.

ConcreteColleague: Implements the Colleague interface. These are the actual objects that need to communicate through the Mediator. They are unaware of each other and communicate only through the Mediator.